Shiza content

+++ src/main/java/com/sbt/task7mskserver/repositories/ClientRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Integer> {
+public interface ClientRepository extends JpaRepository<Client, Long> {
 }
